+++ src/Symfony/Component/Cache/Adapter/AbstractTagAwareAdapter.php
@@ -150,6 +150,16 @@ abstract class AbstractTagAwareAdapter implements TagAwareAdapterInterface, TagA
     abstract protected function doInvalidate(array $tagIds): bool;
 
     /**
+     * Returns the tags bound to the provided ids.
+     */
+    protected function doFetchTags(array $ids): iterable
+    {
+        foreach ($this->doFetch($ids) as $id => $value) {
+            yield $id => \is_array($value) && \is_array($value['tags'] ?? null) ? $value['tags'] : [];
+        }
+    }
+
+    /**
      * {@inheritdoc}
      *
      * @return bool
@@ -232,10 +242,14 @@ abstract class AbstractTagAwareAdapter implements TagAwareAdapterInterface, TagA
             unset($this->deferred[$key]);
         }
 
-        foreach ($this->doFetch($ids) as $id => $value) {
-            foreach ($value['tags'] ?? [] as $tag) {
-                $tagData[$this->getId(self::TAGS_PREFIX.$tag)][] = $id;
+        try {
+            foreach ($this->doFetchTags($ids) as $id => $tags) {
+                foreach ($tags as $tag) {
+                    $tagData[$this->getId(self::TAGS_PREFIX.$tag)][] = $id;
+                }
             }
+        } catch (\Exception $e) {
+            // ignore unserialization failures
         }
 
         try {